login logout register functionality Blocked pages if not logged in

+++ darwaazay/doors/views.py
@@ -1,8 +1,56 @@
 from django.shortcuts import render, redirect
+from django.http import HttpResponse
+from django.contrib import messages
 from django.db.models import Q
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.auth.forms import UserCreationForm
 from . import models
 from . import forms
 # Create your views here.
+
+def loginPage(request):
+    pagename = 'login'
+
+    if request.user.is_authenticated:
+        return redirect('home')
+
+    if request.method == "POST":
+        username = request.POST.get('username').lower()
+        password = request.POST.get('password')
+
+        user = authenticate(request, username=username, password=password)
+
+        if(user):
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, 'Username or password does not exist')
+
+
+    context = {'pagename':pagename}
+    return render(request, 'doors/login_register.html', context)
+
+def logoutUser(request):
+    logout(request)
+    return redirect('home')
+
+def registerUser(request):
+    if request.method == "POST":
+        form = UserCreationForm(request.POST)
+        if form.is_valid():
+            user = form.save(commit=False)
+            user.username = user.username.lower()
+            user.save()
+            login(request, user)
+            return redirect('home')
+        else:
+            messages.error(request, "An error occured during registration")
+
+    pagename = 'register'
+    form = UserCreationForm()
+    context = {'pagename':pagename, 'form':form}
+    return render(request, 'doors/login_register.html', context)
 
 def home(request):
     q = request.GET.get('q')
@@ -25,6 +73,7 @@ def room(request, pk):
     context = {'room':room}
     return render(request, 'doors/room.html', context)
 
+@login_required(login_url="login")
 def createRoom(request):
     if request.method == 'POST':
         form = forms.RoomForm(request.POST)
@@ -35,8 +84,13 @@ def createRoom(request):
     context = {'form':form}
     return render(request, 'doors/room_form.html', context)
 
+@login_required(login_url="login")
 def updateRoom(request, pk):
     room = models.Room.objects.get(id=int(pk))
+
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
+
     if request.method == 'POST':
         form = forms.RoomForm(request.POST, instance=room)
         if form.is_valid():
@@ -46,8 +100,11 @@ def updateRoom(request, pk):
     context = {'form':form}
     return render(request, 'doors/room_form.html', context)
 
+@login_required(login_url="login")
 def deleteRoom(request,pk):
     room = models.Room.objects.get(id=int(pk))
+    if request.user != room.host:
+        return HttpResponse("You are not allowed here")
     if request.method == 'POST':
         room.delete()
         return redirect('home')
